telegram_responder/requirements.txt
python-telegram-bot~=22.0
aiohttp~=3.11.18
schedule~=1.2.2
python-dotenv>=0.19.0
